+++ entity/src/main/java/org/grnet/cat/entities/registry/Motivation.java
@@ -15,6 +15,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.grnet.cat.entities.registry.generator.RegistryId;
 
+import java.sql.Timestamp;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Set;
@@ -62,6 +63,14 @@ public class Motivation extends Registry {
     )
     private Set<MotivationPrincipleJunction> principles = new HashSet<>();
 
+    @OneToMany(
+            mappedBy = "motivation",
+            cascade = CascadeType.ALL,
+            orphanRemoval = true,
+            fetch = FetchType.LAZY
+    )
+    private Set<MotivationActorJunction> actors = new HashSet<>();
+
     public void addPrinciple(Principle principle, String annotationText, String annotationURL, Relation relation, String motivationX, Integer lodMpV) {
 
         var principleMotivation = new MotivationPrincipleJunction(this, principle, annotationText, annotationURL,relation, motivationX, lodMpV);
@@ -82,5 +91,34 @@ public class Motivation extends Registry {
                 motivationPrinciple.setPrinciple(null);
             }
         }
+    }
+    public void addActor(RegistryActor actor, Relation relation, String motivationX, Integer lodMAV,String populatedBy, Timestamp lastTouch) {
+
+        var actorMotivation = new MotivationActorJunction(this,actor,relation, motivationX, lodMAV,populatedBy,lastTouch);
+        actors.add(actorMotivation);
+        actorMotivation.getMotivation().actors.add(actorMotivation);
+    }
+//
+//    public void removeActor(RegistryActor actor) {
+//        for (Iterator<MotivationActorJunction> iterator = actors.iterator();
+//             iterator.hasNext(); ) {
+//            var motivationActor = iterator.next();
+//
+//            if (motivationActor.getMotivation().equals(this) &&
+//                    motivationActor.getActor().equals(actor)) {
+//                iterator.remove();
+//                motivationActor.getActor().getMotivations().remove(motivationActor);
+//                motivationActor.setMotivation(null);
+//                motivationActor.setActor(null);
+//            }
+//        }
+//    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
     }
 }