+++ CHANGELOG.md
@@ -22,6 +22,10 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 
 ## Unreleased
 
+---
+### Added
+
+- [#269](https://github.com/FC4E-CAT/fc4e-cat-api/pull/269) CAT-476 Motivations and actors (motivation-actor relation)
 
 ## 1.7.0 - 2024-09-10